Moves fragments
