+++ app/src/main/res/layout/fragment_moves.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="com.swanberg.pogoiv.debug.MovesFragment">
+    tools:context=".Fragments.MovesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
